Handle HTTPS -> HTTP downgrades in bearer tokens
Currently, Library Simplified uses HTTPS. It may serve bearer tokens
that point to servers hosted by publishers that run on HTTP. If those
publisher's servers return a redirect, `okhttp` will refuse to follow
the redirect because it considers the action to be an HTTPS -> HTTP
downgrade. As we know this is specifically going to happen with bearer
tokens, we add an explicit workaround in the bearer token interceptor
that makes a sanitized request to the HTTP server.